pipelines/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
numpy>=1.20.0
scipy>=1.7.0
Pillow>=8.0.0
matplotlib>=3.4.0
imageio>=2.9.0
trimesh>=3.9.0
scikit-image>=0.18.0
tqdm>=4.62.0
facenet-pytorch>=2.5.0
h5py>=3.1.0
kornia>=0.6.0
timm>=0.4.12
trimesh>=3.9.0
open3d>=0.13.0
scikit-learn>=0.24.0
seaborn>=0.11.0